bundles/co2mon/files/requirements.txt
requests
paho-mqtt
git+https://github.com/heinemml/CO2Meter
